arp: support a proxy ARP IP address on the netif
allows (for example) for a PPP <-> ethernet gateway to support proxy
ARP. This is simpler than the (currently disabled) proxyarp support
built into PPP in lwip and more general purpose.

To use this the application should call netif_set_proxyarp() on the
interface that should present the extra IP address, this will usually
be a primary ethernet interface. When used with PPP the IP passed in
should be the IP address of the remote end of the PPP link.

This feature is enabled with the ARP_PROXYARP_SUPPORT define.