pppos: fixed reservation for forwarding with VJ compression
when the frame is VJ compressed we need to reserve space for the
ethernet header if forwarding is enabled. The symptom this fixes is
that the first data frame after a TCP connection is established can be
lost as there is no room to add in the ethernet header in the pbuf,
and re-sends hit the same issue, resulting in the TCP connection never
sending the first block of data and thus stalling.